further refactor predict.py (#944)
* new case studies

* tests not yet passing

* omitted files

* don't subtract 8

* more tests passing

* two tests skipped, rest working

* clear notebook

* copy the cfg

* moved mock_test to api case study

* removed redundant code

* import case_studies version of mock_tests

* missed variable name changes

* unsaved substitutions

* moved igs.py file to api case study; remove unused Predict tests

* remove no-long-used crop method

* pylint

* add test coverage of CachedSimulatedDataset and train.py

* use CachedSimulatedDataset in test
