+++ .flake8
@@ -124,5 +124,8 @@ ignore =
     # what's wrong a multiline try block?
     WPS229,
 
+    # lambda expressions are more concise and very readable
+    E731,
+
     # we should consider fixing some of these issues
     W503,WPS202,WPS213,WPS214,WPS231,WPS236,WPS336,WPS338,WPS440,WPS441,WPS442,WPS602
