+++ bliss/predict.py
@@ -1,378 +1,23 @@
-from pathlib import Path
-
-import numpy as np
 import torch
-from bokeh.models import TabPanel, Tabs
-from bokeh.plotting import figure, output_file, show
 from hydra.utils import instantiate
-from reproject import reproject_interp
-from torch import Tensor
-
-from bliss.catalog import FullCatalog
-from bliss.surveys.sdss import PhotoFullCatalog
-from bliss.surveys.sdss import SloanDigitalSkySurvey as SDSS
 
 
-def crop_image(x, cfg):
-    """Crop the image to a subregion for prediction."""
-    # image dimensions must be a multiple of 16
-    height = x.size(2) - (x.size(2) % 16)
-    width = x.size(3) - (x.size(3) % 16)
-    x = x[:, :, :height, :width].float()
-
-    cp = cfg.predict.crop
-    if cp.do_crop:
-        top_left_y, top_left_x = cp.left_upper_corner
-        w, h = cp.width, cp.height
-        if ((top_left_y + h) <= x.shape[2]) and ((top_left_x + w) <= x.shape[3]):
-            x = x[:, :, top_left_y : top_left_y + h, top_left_x : top_left_x + w]
-
-    return x
-
-
-def align(img, wcs_list, ref_band, ref_depth=0):
-    """Reproject images based on some reference WCS for pixel alignment."""
-    reproj_d = {}
-    footprint_d = {}
-
-    # coerce to 4D with coadd_depth trivially 1
-    if img.ndim == 3:
-        img = np.expand_dims(img, axis=0)
-    if not isinstance(wcs_list[0], list):
-        wcs_list = [wcs_list]
-    coadd_depth, n_bands, h, w = img.shape
-
-    # align with (`ref_depth`, `ref_band`) WCS
-    target_wcs = wcs_list[ref_depth][ref_band]
-    for d in range(coadd_depth):
-        for bnd in range(n_bands):
-            inputs = (img[d, bnd], wcs_list[d][bnd])
-            # the next line is the computational bottleneck
-            reproj, footprint = reproject_interp(
-                inputs, target_wcs, order="bicubic", shape_out=(h, w)
-            )
-            reproj_d[(d, bnd)] = reproj
-            footprint_d[(d, bnd)] = footprint
-
-    # use footprints to handle NaNs from reprojection
-    fp_h, fp_w = footprint_d[(0, 0)].shape
-    out_print = np.ones((fp_h, fp_w))
-    for fp in footprint_d.values():
-        out_print *= fp
-
-    out_print = np.expand_dims(out_print, axis=(0, 1))
-    reproj_out = np.zeros(img.shape)
-    for d in range(coadd_depth):
-        for bnd in range(n_bands):
-            reproj_d[(d, bnd)] = np.multiply(reproj_d[(d, bnd)], out_print)
-            cropped = reproj_d[(d, bnd)][0, 0, :h, :w]
-            cropped[np.isnan(cropped)] = 0
-            reproj_out[(d, bnd)] = cropped
-
-    if reproj_out.shape[0] == 1:
-        reproj_out = reproj_out.squeeze(axis=0)
-    return reproj_out
-
-
-def nelec_to_nmgy_for_catalog(est_cat, nelec_per_nmgy_per_band):
-    fluxes_suffix = "_fluxes"
-    # reshape nelec_per_nmgy_per_band to (1, 1, 1, 1, {n_bands}) to broadcast
-    nelec_per_nmgy_per_band = torch.tensor(nelec_per_nmgy_per_band, device=est_cat.device)
-    nelec_per_nmgy_per_band = nelec_per_nmgy_per_band.view(1, 1, 1, 1, -1)
-    for key in est_cat.keys():
-        if key.endswith(fluxes_suffix):
-            est_cat[key] = est_cat[key] / nelec_per_nmgy_per_band
-    return est_cat
-
-
-def predict_frame(cfg, frame, encoder):
-    aligned_images = align(
-        frame["image"],
-        wcs_list=frame["wcs"],
-        ref_band=cfg.simulator.prior.reference_band,
-    )
-    aligned_backgrounds = align(
-        frame["background"],
-        wcs_list=frame["wcs"],
-        ref_band=cfg.simulator.prior.reference_band,
-    )
-
-    device = encoder.device
-    images = torch.from_numpy(aligned_images).unsqueeze(0).to(device=device)
-    backgrounds = torch.from_numpy(aligned_backgrounds).unsqueeze(0).to(device=device)
-
-    # cropping the images should be done by the dataset
-    images = crop_image(images, cfg)
-    backgrounds = crop_image(backgrounds, cfg)
-
-    batch = {"images": images, "background": backgrounds}
-    est_cat, pred = encoder.predict_step(batch, None).values()
-
-    # mean of the nelec_per_mgy per band
-    nelec_per_nmgy_per_band = np.mean(frame["flux_calibration_list"], axis=1)
-    est_cat = nelec_to_nmgy_for_catalog(est_cat, nelec_per_nmgy_per_band)
-    est_full = est_cat.to_full_params()
-
-    return est_full, pred, images.squeeze(0), backgrounds.squeeze(0)
-
-
-def predict_and_compare(cfg):
-    dataset = instantiate(cfg.predict.dataset, load_image_data=True)
-
-    images_for_frame = {}
-    radecs_for_frame = {}
-    backgrounds_for_frame = {}
-    preds_for_frame = {}
-
-    # collated catalog for all images
-    survey_plocs_all = None
-    bliss_cat_all = None
-
-    encoder = instantiate(cfg.encoder).to(cfg.predict.device)
-    enc_state_dict = torch.load(cfg.predict.weight_save_path)
-    encoder.load_state_dict(enc_state_dict)
+def predict(cfg):
+    encoder = instantiate(cfg.encoder)
     encoder.eval()
+    encoder.load_state_dict(torch.load(cfg.predict.weight_save_path))
 
-    for i, frame in enumerate(dataset):
-        images_id = dataset.image_id(i)  # should be called frame_id, not image_id
+    dataset = instantiate(cfg.predict.dataset)
 
-        bliss_cat, bliss_pred, images, backgrounds = predict_frame(cfg, frame, encoder)
-        images_for_frame[images_id] = images
-        backgrounds_for_frame[images_id] = backgrounds
-        preds_for_frame[images_id] = bliss_pred
+    trainer = instantiate(cfg.training.trainer)
+    enc_output = trainer.predict(encoder, datamodule=dataset)
 
-        # we are merging bliss catalogs as we go here, but we shouldn't be!
-        # we should keep each bliss_cat associated with a particular frame_id
-        if bliss_cat_all is None:
-            bliss_cat_all = bliss_cat
-        else:
-            d = {}
-            d["plocs"] = torch.cat((bliss_cat_all.plocs, bliss_cat.plocs), dim=1)
-            d["n_sources"] = Tensor([bliss_cat_all.n_sources + bliss_cat.n_sources])
-            est_full_all_dict = bliss_cat_all.to_dict()
-            for k, v in bliss_cat.items():
-                d[k] = torch.cat((est_full_all_dict[k], v), dim=1)
-            bliss_cat_all = FullCatalog(bliss_cat_all.height, bliss_cat_all.width, d)
+    est_cat_tables = []
+    for batch_output in enc_output:
+        est_cat = batch_output["est_cat"]
+        full_cat = est_cat.to_full_catalog()
+        astropy_cat = full_cat.to_astropy_table(cfg.encoder.survey_bands)
+        est_cat_tables.append(astropy_cat)
 
-        # store the RA and Dec for the center of the image
-        ra, dec = frame["wcs"][cfg.simulator.prior.reference_band].all_pix2world(
-            frame["image"].shape[2] / 2, frame["image"].shape[1] / 2, 0
-        )
-        radecs_for_frame[images_id] = (ra.item(), dec.item())
-
-        # now we load the survey's catalog from this frame
-        survey_cat_path = dataset.downloader.download_catalog(images_id)
-        try:
-            survey_cat = dataset.catalog_cls.from_file(
-                cat_path=survey_cat_path,
-                wcs=frame["wcs"][cfg.simulator.prior.reference_band],
-                height=frame["image"].shape[1],
-                width=frame["image"].shape[2],
-            )
-            survey_plocs = survey_cat.plocs[0]
-        except IndexError:
-            survey_plocs = None
-
-        # we are merging survey catalogs as we go here, but we shouldn't be!
-        # we should keep each catalog associated with a particular frame_id
-        if survey_plocs_all is None:
-            survey_plocs_all = survey_plocs
-        else:
-            survey_plocs_all = torch.cat((survey_plocs_all, survey_plocs), dim=0)
-            survey_plocs_all = torch.unique(survey_plocs_all, dim=0)
-
-    # better not to have this here...the caller should call plot_predict directly
-    if cfg.predict.plot.show_plot and (survey_plocs_all is not None):
-        plot_predict(
-            cfg,
-            images_for_frame,
-            backgrounds_for_frame,
-            radecs_for_frame,
-            survey_plocs_all,
-            bliss_cat_all,
-        )
-
-    return bliss_cat_all, images_for_frame, backgrounds_for_frame, survey_plocs_all, preds_for_frame
-
-
-def crop_plocs(cfg, w, h, plocs, do_crop=False):
-    tiles_to_crop = cfg.encoder.tiles_to_crop
-    tile_slen = cfg.encoder.tile_slen
-    minh = tiles_to_crop * tile_slen
-    maxh = h + tiles_to_crop * tile_slen
-    minw = tiles_to_crop * tile_slen
-    maxw = w + tiles_to_crop * tile_slen
-    if do_crop:
-        minh += cfg.predict.crop.left_upper_corner[0]
-        maxh += cfg.predict.crop.left_upper_corner[0]
-        minw += cfg.predict.crop.left_upper_corner[1]
-        maxw += cfg.predict.crop.left_upper_corner[1]
-
-    x0_mask = (plocs[:, 0] > minh) & (plocs[:, 0] < maxh)
-    x1_mask = (plocs[:, 1] > minw) & (plocs[:, 1] < maxw)
-    x_mask = x0_mask * x1_mask
-
-    return plocs[x_mask].cpu() - torch.tensor([minh, minw])
-
-
-def add_cat(p, est_plocs, survey_true_plocs, sdss_plocs, decals_plocs):
-    """Function to overlay scatter plots on given image using different catalogs."""
-    p.scatter(
-        est_plocs[:, 1],
-        est_plocs[:, 0],
-        color="darksalmon",
-        marker="circle",
-        legend_label="estimated catalog",
-        size=10,
-        fill_color=None,
-    )
-    if survey_true_plocs is not None:
-        p.scatter(
-            survey_true_plocs[:, 1],
-            survey_true_plocs[:, 0],
-            marker="circle",
-            color="hotpink",
-            legend_label="consolidated survey catalog",
-            size=20,
-            fill_color=None,
-        )
-    if sdss_plocs is not None:
-        p.scatter(
-            sdss_plocs[:, 1],
-            sdss_plocs[:, 0],
-            marker="circle",
-            color="lightgreen",
-            legend_label="sdss catalog",
-            size=5,
-            fill_color=None,
-        )
-    if decals_plocs is not None:
-        p.scatter(
-            decals_plocs[:, 1],
-            decals_plocs[:, 0],
-            marker="circle",
-            color="cyan",
-            legend_label="decals catalog",
-            size=5,
-        )
-    p.legend.click_policy = "hide"
-    return p
-
-
-def plot_image(cfg, ra, dec, img, w, h, est_plocs, survey_true_plocs, title):
-    """Function that generate plots for images."""
-    from bliss.surveys.decals import DECaLS  # pylint: disable=import-outside-toplevel
-    from bliss.surveys.des import TractorFullCatalog  # pylint: disable=import-outside-toplevel
-
-    p = figure(width=cfg.predict.plot.width, height=cfg.predict.plot.height)
-    p.image(image=[img], x=0, y=0, dw=w, dh=h, palette="Viridis256")
-    do_crop = cfg.predict.crop.do_crop
-    is_simulated = cfg.predict.is_simulated
-
-    survey_true_plocs = np.array(crop_plocs(cfg, w, h, survey_true_plocs, do_crop).cpu())
-    sdss_plocs = None
-    decals_plocs = None
-    if do_crop and (not is_simulated):
-        # DECaLS one-instance catalog plocs
-        decals = instantiate(
-            cfg.surveys.decals, bands=[DECaLS.BANDS.index("r")], sky_coords=[{"ra": ra, "dec": dec}]
-        )
-        brickname = DECaLS.brick_for_radec(ra, dec)
-        tractor_filename = decals.downloader.download_catalog(brickname)
-        decals_plocs = TractorFullCatalog.from_file(
-            tractor_filename,
-            wcs=decals[0]["wcs"][cfg.simulator.prior.reference_band],
-            height=decals[0]["background"].shape[1],
-            width=decals[0]["background"].shape[2],
-        ).plocs[0]
-        decals_plocs = np.array(crop_plocs(cfg, w, h, decals_plocs, do_crop).cpu())
-
-        # SDSS one-instance catalog plocs
-        run, camcol, field = SDSS.rcf_for_radec(ra, dec)
-        sdss = instantiate(
-            cfg.surveys.sdss, fields=[{"run": run, "camcol": camcol, "fields": [field]}]
-        )
-        photocat_filename = sdss.downloader.download_catalog((run, camcol, field))
-        sdss_plocs = PhotoFullCatalog.from_file(
-            photocat_filename,
-            wcs=sdss[0]["wcs"][cfg.simulator.prior.reference_band],
-            height=sdss[0]["background"].shape[1],
-            width=sdss[0]["background"].shape[2],
-        ).plocs[0]
-        sdss_plocs = np.array(crop_plocs(cfg, w, h, sdss_plocs, do_crop).cpu())
-    return TabPanel(
-        child=add_cat(p, est_plocs, survey_true_plocs, sdss_plocs, decals_plocs), title=title
-    )
-
-
-def plot_predict(
-    cfg,
-    images_for_frame,
-    backgrounds_for_frame,
-    radecs_for_frame,
-    survey_true_plocs,
-    est_cat: FullCatalog,
-):
-    """Function that uses bokeh to save generated plots to an html file."""
-    if cfg.predict.plot.out_file_name is not None:
-        out_filepath = Path(cfg.predict.plot.out_file_name)
-        out_filepath.parent.mkdir(parents=True, exist_ok=True)
-        output_file(out_filepath)
-
-    est_plocs = np.array(est_cat.plocs.cpu())[0]
-    est_tile = est_cat.to_tile_params(
-        cfg.encoder.tile_slen,
-        cfg.simulator.prior.max_sources,
-        ignore_extra_sources=True,
-    ).to("cpu")
-
-    tabs = []
-    for image_id in images_for_frame.keys():
-        image = images_for_frame[image_id]
-        background = backgrounds_for_frame[image_id]
-
-        ptc = cfg.encoder.tiles_to_crop * cfg.encoder.tile_slen
-        image = image[0, ptc:-ptc, ptc:-ptc]  # uh, are we always plotting the u band image here?
-        background = background[0, ptc:-ptc, ptc:-ptc]
-
-        w, h = image.shape
-
-        ra, dec = radecs_for_frame[image_id]
-        run, camcol, field = SDSS.rcf_for_radec(ra, dec)
-        simulator = instantiate(
-            cfg.simulator,
-            survey={"fields": [{"run": run, "camcol": camcol, "fields": [field]}]},
-        )
-        decoder_obj = simulator.image_decoder
-        recon_images = decoder_obj.render_images(est_tile, [(run, camcol, field)])[
-            0
-        ]  # NOTE: causes issue when weights used were for survey that has diff no. of bands as SDSS
-        recon_img = recon_images[0][0]  # first image in batch, first band in image
-
-        image = image.to("cpu")
-        background = background.to("cpu")
-        res = image - recon_img - background
-
-        # normalize for big image
-        title_suffix = f" (RA: {float(ra):.2f}, Dec: {float(dec):.2f})"
-        title_prefix = ""
-        if w >= 150:
-            image = np.log((image - image.min()) + 10)
-            recon_img = np.log((recon_img - recon_img.min()) + 10)
-            title_prefix = "log-"
-
-        np_image = np.array(image)
-        np_recon = np.array(recon_img)
-        np_res = np.array(res)
-        tab1_title = f"{title_prefix}true image{title_suffix}"
-        tab2_title = f"{title_prefix}reconstructed image{title_suffix}"
-        tab3_title = f"residual{title_suffix}"
-
-        # we are plotting detections from all frames on each frame here, but we shouldn't be!
-        # each frame has it's own bliss catalog and it's own survey catalog
-        tab1 = plot_image(cfg, ra, dec, np_image, w, h, est_plocs, survey_true_plocs, tab1_title)
-        tab2 = plot_image(cfg, ra, dec, np_recon, w, h, est_plocs, survey_true_plocs, tab2_title)
-        tab3 = plot_image(cfg, ra, dec, np_res, w, h, est_plocs, survey_true_plocs, tab3_title)
-        tabs.extend([tab1, tab2, tab3])
-
-    show(Tabs(tabs=tabs))
+    # eventually we should return variational distribution samples too
+    return est_cat_tables
