+++ bliss/simulator/simulated_dataset.py
@@ -10,9 +10,9 @@ from torch import Tensor
 from torch.utils.data import DataLoader, Dataset, IterableDataset
 from tqdm import tqdm
 
+from bliss.align import align
 from bliss.catalog import TileCatalog
 from bliss.generate import FileDatum
-from bliss.predict import align
 from bliss.simulator.decoder import ImageDecoder
 from bliss.simulator.prior import CatalogPrior
 from bliss.surveys.survey import Survey
@@ -38,6 +38,8 @@ class SimulatedDataset(pl.LightningDataModule, IterableDataset):
         super().__init__()
 
         self.survey = survey
+        survey.prepare_data()
+
         self.catalog_prior = prior
         self.background = self.survey.background
         assert self.catalog_prior is not None, "Survey prior cannot be None."
@@ -100,14 +102,13 @@ class SimulatedDataset(pl.LightningDataModule, IterableDataset):
         """Align images to the reference depth and band."""
         batch_size = images.shape[0]
         for b in range(batch_size):
-            images[b] = torch.from_numpy(
-                align(
-                    images[b].numpy(),
-                    wcs_list=wcs_batch[b],
-                    ref_depth=0,
-                    ref_band=self.catalog_prior.reference_band,
-                )
+            aligned_image = align(
+                images[b].numpy(),
+                wcs_list=wcs_batch[b],
+                ref_depth=0,
+                ref_band=self.catalog_prior.reference_band,
             )
+            images[b] = torch.from_numpy(aligned_image)
         return images
 
     def simulate_image(
