+++ bliss/surveys/des.py
@@ -11,11 +11,9 @@ from astropy.table import Table
 from astropy.wcs import WCS, FITSFixedWarning
 from numpy.core import defchararray
 from omegaconf import DictConfig
-from pytorch_lightning.utilities.types import EVAL_DATALOADERS
 from pyvo.dal import sia
 from scipy.interpolate import RectBivariateSpline
 from scipy.ndimage import zoom
-from torch.utils.data import DataLoader
 
 from bliss.catalog import FullCatalog, SourceType
 from bliss.simulator.background import ImageBackground
@@ -264,23 +262,6 @@ class DarkEnergySurvey(Survey):
                 im_id[b] = im_id.get(b, "")
         # convert to hashable DictConfig
         return [self.to_dictconfig(im_id) for im_id in im_ids]
-
-    @property
-    def predict_batch(self):
-        if not self._predict_batch:
-            self._predict_batch = {
-                "images": self[0]["image"],
-                "background": self[0]["background"],
-            }
-        return self._predict_batch
-
-    @predict_batch.setter
-    def predict_batch(self, value):
-        self._predict_batch = value
-
-    def predict_dataloader(self) -> EVAL_DATALOADERS:
-        assert self.predict_batch is not None, "predict_batch must be set."
-        return DataLoader([self.predict_batch], batch_size=1)
 
 
 DES = DarkEnergySurvey
