+++ case_studies/dependent_tiling/config.yaml
@@ -49,18 +49,14 @@ predict:
             pixel_scale: 0.396
             psf_slen: 25
         pixel_shift: 2
+        align_to_band: null
+        crop_config:
+            do_rectangle_crop: true
+            left_upper_corner: [630, 310]
+            width: 112
+            height: 112
     trainer: ${training.trainer}
     encoder: ${encoder}
     weight_save_path: ${paths.root}/case_studies/dependent_tiling/m2.pt
     device: "cuda:0"
-    crop:
-        do_crop: true
-        left_upper_corner: [630, 310]
-        width: 112
-        height: 112
-    plot:
-        show_plot: true
-        width: 1000
-        height: 1000
-        out_file_name: ${paths.output}/predict.html
     is_simulated: false
