+++ docs/docsrc/conf.py
@@ -13,7 +13,6 @@ import os
 import sys
 
 sys.path.insert(0, os.path.abspath("../../bliss"))
-print(sys.path)
 
 
 # -- Project information -----------------------------------------------------
