+++ tests/test_catalogs.py
@@ -1,4 +1,3 @@
-import warnings
 from pathlib import Path
 
 import numpy as np
@@ -9,7 +8,6 @@ from hydra.utils import instantiate
 from bliss.catalog import FullCatalog, SourceType, TileCatalog
 from bliss.surveys.decals import DarkEnergyCameraLegacySurvey as DECaLS
 from bliss.surveys.decals import TractorFullCatalog
-from case_studies.adaptive_tiling.region_catalog import RegionCatalog, tile_cat_to_region_cat
 
 # TODO: Add PhotoFullCatalog-specific tests (like loading, restricting by RA/DEC, downloading)
 
@@ -49,38 +47,6 @@ def multi_source_tilecat():
     return TileCatalog(4, d)
 
 
-@pytest.fixture(scope="module")
-def region_cat():
-    n_sources = torch.zeros(2, 5, 5)
-    n_sources[0, 0, 0] = 1
-    n_sources[0, 1, 3] = 1
-    n_sources[0, 4, 2] = 1
-    n_sources[1, 0, 0] = 1
-    n_sources[1, 2, 2] = 1
-    n_sources[1, 4, 4] = 1
-
-    locs = torch.zeros(2, 5, 5, 1, 2)
-    locs[0, 0, 0] = torch.tensor([0.8, 0.2])
-    locs[0, 1, 3] = torch.tensor([0.1, 0.5])
-    locs[0, 4, 2] = torch.tensor([0.3, 0.3])
-    locs[1, 0, 0] = torch.tensor([0.5, 0.5])
-    locs[1, 2, 2] = torch.tensor([0.5, 0.5])
-    locs[1, 4, 4] = torch.tensor([0.5, 0.5])
-
-    fluxes = torch.zeros(2, 5, 5, 1, 5)
-    fluxes[0, 0, 0] = torch.tensor([100, 100, 100, 100, 100])
-    fluxes[0, 1, 3] = torch.tensor([100, 100, 100, 100, 100]) * 2
-    fluxes[0, 4, 2] = torch.tensor([100, 100, 100, 100, 100]) * 5
-
-    d = {
-        "n_sources": n_sources,
-        "locs": locs,
-        "galaxy_fluxes": fluxes,
-    }
-
-    return RegionCatalog(interior_slen=3.5, overlap_slen=0.5, d=d)
-
-
 class TestBasicTileAndFullCatalogs:
     def test_unallowed_param(self):
         d_tile = {
@@ -114,7 +80,7 @@ class TestBasicTileAndFullCatalogs:
         assert "source_type" in keys
         assert "galaxy_bools" not in keys
 
-        full_cat = tile_cat.to_full_params()
+        full_cat = tile_cat.to_full_catalog()
         assert full_cat.plocs.equal(full_cat["plocs"])
         assert full_cat.galaxy_bools.equal(full_cat["galaxy_bools"])
 
@@ -160,15 +126,15 @@ class TestBasicTileAndFullCatalogs:
         full_cat = FullCatalog(2, 2, d)
 
         with pytest.raises(ValueError) as error_info:
-            full_cat.to_tile_params(1, 1, ignore_extra_sources=False)
+            full_cat.to_tile_catalog(1, 1, ignore_extra_sources=False)
         assert error_info.value.args[0] == "# of sources per tile exceeds `max_sources_per_tile`."
 
         # should return only first source in first tile.
-        tile_cat = full_cat.to_tile_params(1, 1, ignore_extra_sources=True)
+        tile_cat = full_cat.to_tile_catalog(1, 1, ignore_extra_sources=True)
         assert torch.equal(tile_cat.n_sources, torch.tensor([[[1, 0], [0, 0]]]))
 
         # test to_tile_coords and to_full_coords (set max_sources_per_tile to 2)
-        convert_full_cat = full_cat.to_tile_params(1, 2).to_full_params()
+        convert_full_cat = full_cat.to_tile_catalog(1, 2).to_full_catalog()
         assert torch.allclose(convert_full_cat.plocs, full_cat.plocs)
 
         correct_locs = torch.tensor([[[0.5, 0.5], [0, 0]], [[0, 0], [0, 0]]]).reshape(1, 2, 2, 1, 2)
@@ -204,84 +170,3 @@ class TestDecalsCatalog:
         assert np.isclose(np.max(ras), 336.75, atol=1e-4)
         assert np.isclose(np.min(decs), -1.125, atol=1e-4)
         assert np.isclose(np.max(decs), -0.875, atol=1e-4)
-
-
-class TestRegionCatalog:
-    def test_properties(self, region_cat):
-        assert region_cat.height == region_cat.width == 12
-        assert region_cat.is_on_mask.sum() == 6
-        assert torch.all(
-            region_cat.interior_mask + region_cat.boundary_mask + region_cat.corner_mask,
-        )
-        assert not torch.any(
-            region_cat.interior_mask * region_cat.boundary_mask * region_cat.corner_mask,
-        )
-
-    def test_region_coords(self, region_cat):
-        coords = region_cat.get_region_coords()
-        assert coords.amax(dim=(0, 1))[0] < region_cat.height
-        assert coords.amax(dim=(0, 1))[1] < region_cat.width
-
-    def test_region_sizes(self, region_cat):
-        sizes = region_cat.get_region_sizes()
-        assert sizes[0].equal(
-            torch.tensor([[3.75, 3.75], [3.75, 0.5], [3.75, 3.5], [3.75, 0.5], [3.75, 3.75]])
-        )
-        assert sizes[1].equal(
-            torch.tensor([[0.5, 3.75], [0.5, 0.5], [0.5, 3.5], [0.5, 0.5], [0.5, 3.75]])
-        )
-        assert torch.all(sizes[..., 0].sum(dim=0) == region_cat.height)
-        assert torch.all(sizes[..., 1].sum(dim=1) == region_cat.width)
-
-    def test_convert_to_full(self, region_cat):
-        full_cat = region_cat.to_full_params()
-        true_locs = torch.tensor(
-            [
-                [[9.375, 5.3], [3.8, 8], [3, 0.75]],
-                [[6, 6], [10.125, 10.125], [1.875, 1.875]],
-            ]
-        )
-        assert full_cat.plocs.equal(true_locs)
-
-    def test_tile_cat_to_region_basic(self, basic_tilecat):
-        region_cat = tile_cat_to_region_cat(basic_tilecat, 0.5, discard_extra_sources=False)
-        full_cat = basic_tilecat.to_full_params()
-        assert region_cat.to_full_params().plocs.equal(full_cat.plocs)
-
-    def test_tile_cat_to_region_filtering(self):
-        d = {
-            "n_sources": torch.zeros(3, 2, 2),
-            "locs": torch.zeros(3, 2, 2, 1, 2),
-            "source_type": torch.ones((3, 2, 2, 1, 1)).bool(),
-            "galaxy_params": torch.zeros((3, 2, 2, 1, 6)),
-            "star_fluxes": torch.ones((3, 2, 2, 1, 5)) * 1000,
-            "galaxy_fluxes": torch.ones(3, 2, 2, 1, 5) * 1000,
-        }
-        # BATCH 0: top right interior, center right boundary
-        d["n_sources"][0, 0, 1] = 1
-        d["n_sources"][0, 1, 1] = 1
-        d["locs"][0, 0, 1, 0] = torch.tensor([0.5, 0.5])
-        d["locs"][0, 1, 1, 0] = torch.tensor([0.02, 0.5])
-        d["galaxy_fluxes"][0, 0, 1, 0, 2] = 5000  # keep top right
-
-        # BATCH 1: top left interior, top center boundary
-        d["n_sources"][1, 0, 0] = 1
-        d["n_sources"][1, 0, 1] = 1
-        d["locs"][1, 0, 0, 0] = torch.tensor([0.5, 0.5])
-        d["locs"][1, 0, 1, 0] = torch.tensor([0.5, 0.02])
-        d["galaxy_fluxes"][1, 0, 1, 0, 2] = 5000  # keep top center
-
-        # BATCH 2: only one source in top left
-        d["n_sources"][2, 0, 0] = 1
-        d["locs"][2, 0, 0, 0] = torch.tensor([0.5, 0.5])
-
-        tilecat = TileCatalog(4, d)
-
-        # make sure no warning when converting (since extra sources have been discarded)
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            region_cat = tile_cat_to_region_cat(tilecat, 0.5, discard_extra_sources=True)
-
-        n_sources = region_cat.n_sources
-        assert n_sources[0, 0, 2] == n_sources[1, 0, 1] == n_sources[2, 0, 0] == 1
-        assert n_sources[0].sum() == n_sources[1].sum() == n_sources[2].sum()
