+++ tests/test_sdss.py
@@ -10,6 +10,7 @@ class TestSDSS:
         the_cfg = cfg.copy()
         the_cfg.surveys.sdss.fields = [{"run": 3900, "camcol": 6, "fields": [269]}]
         sdss_obj = instantiate(the_cfg.surveys.sdss)
+        sdss_obj.prepare_data()
         an_obj = sdss_obj[0]
         for k in ("background", "gain", "flux_calibration_list", "calibration"):
             assert isinstance(an_obj[k], np.ndarray)
@@ -24,5 +25,7 @@ class TestSDSS:
         if cfg.surveys.sdss.dir_path != the_cfg.surveys.sdss.dir_path:
             copytree(cfg.surveys.sdss.dir_path, the_cfg.surveys.sdss.dir_path, dirs_exist_ok=True)
         # Also tests images loaded
-        sdss_obj = instantiate(the_cfg.surveys.sdss, load_image_data=True)[0]
-        assert sdss_obj["image"].shape == (5, 1489, 2048)
+        sdss_obj = instantiate(the_cfg.surveys.sdss, load_image_data=True)
+        sdss_obj.prepare_data()
+        frame0 = sdss_obj[0]
+        assert frame0["image"].shape == (5, 1488, 2048)
